Fix a doubled counting for features activated via query string.
If a feature were activated via query string and the form is then reloaded, the amount is counted two times.
This is because the reloaded form already has the checkbox checked and so it is counted the first time.
When, then, the script checks for features activated via query string, again does the counting adding the feature amount again to the total.

This results in the amount summed two times.

Adding an `if` fixes the problem.